Allow challenges for more complex FQDNs
Some of ACME challenges may use full-qualified domain names
which are not only consisting of '_acme_challenge.domain.tld.'.
For this cases, we assume the last 2 segments (the trailing
"." being ignored) to be the domain name and everything
before will be used for the DNS record to be updated.

Thanks to @DerKnerd for pointing out the issue and the initial
implementation.

Fixes #1.