feat(chart): allow to specify image hash instead of tag
Instead of using image tags and in order to make always
reproducible builds, an image hash (digest) could be specified
now in webhook chart. This is especially useful in situations,
where you don't want the official image to be deployed, but
a testing one (i used this often in my test scenarios).